+++ src/shared/ui/mods/com.pa.legion-expansion/new_game.js
@@ -9,36 +9,6 @@ function legionNewGame() {
   try {
     var legionExpansionEnabled = false;
 
-    model.legionClientModLoaded = ko.observable(false);
-
-    model.legionDoNotShowWelcome = ko
-      .observable(false)
-      .extend({ local: "legion_welcome_dontshow" });
-
-    model.legionToggleDoNotShowWelcome = function () {
-      model.legionDoNotShowWelcome(!model.legionDoNotShowWelcome());
-    };
-
-    model.legionUrlClicked = function (data, event) {
-      if (_.has(event, "target.href")) {
-        model.legionOpenUrl(event.target.href);
-      }
-    };
-
-    model.legionOpenUrl = function (url) {
-      engine.call("web.launchPage", url);
-    };
-
-    model.legionCloseWelcome = function () {
-      $("#legion-welcome").fadeOut();
-      $("body").off("keypress", model.legionCloseWelcome);
-    };
-
-    model.legionShowWelcome = function () {
-      $("body").on("keypress", model.legionCloseWelcome);
-      $("#legion-welcome").delay(1000).fadeIn();
-    };
-
     model.isLegion = function (commander) {
       // eslint-disable-next-line no-undef
       return _.includes(legion.commanders, commander);
@@ -61,41 +31,6 @@ function legionNewGame() {
       loadCSS(
         "coui://ui/mods/com.pa.legion-expansion/css/legion_commander_picker.css"
       );
-
-      //legion welcome screen
-      loadCSS("coui://ui/mods/com.pa.legion-expansion/css/welcome.css");
-      $("body").append(
-        loadHtml("coui://ui/mods/com.pa.legion-expansion/new_game/welcome.html")
-      );
-
-      api.mods.getMounted("client").then(function (mods) {
-        var legionClientLoaded =
-          _.intersection(_.pluck(mods, "identifier"), [
-            "com.pa.legion-expansion-client",
-            "com.pa.legion-expansion-client-dev",
-          ]).length > 0;
-
-        model.legionClientModLoaded(legionClientLoaded);
-
-        if (!legionClientLoaded) {
-          if (model.registerHoldReady) {
-            model.registerHoldReady(
-              "com.pa.legion-expansion-client",
-              "Legion Client Mod Missing"
-            );
-          }
-          if (model.localChatMessage) {
-            model.localChatMessage(
-              "Legion Expansion",
-              "Legion Expansion client mod is not installed!"
-            );
-          }
-        }
-
-        if (!model.legionDoNotShowWelcome() && !model.returnFromLoad()) {
-          model.legionShowWelcome();
-        }
-      });
 
       loadScript("coui://ui/mods/com.pa.legion-expansion/common.js");